+++ packages/react-components/lib/tasks/descriptions/index.ts
@@ -1,3 +1,5 @@
 // export * from './clean';
+export * from './custom-compose';
 export * from './delivery_custom';
+export * from './patrol';
 export * from './utils';
